refactor: move PublicIpDiscoveryService from core to services/ip
The IPv4 discovery implementation was the only reason `core` depended on
JAX-RS.
Moving it to `services/ip` alongside `PublicIpv6DiscoveryServiceImpl` keeps
all HTTP-based IP-discovery code in one dedicated module.

Changes:
- Rename `PublicIpDiscoveryService` →  `PublicIpv4DiscoveryServiceImpl`;
  move from `core.sip` package to `services.ip` package.
  The class now implements `PublicIpv4DiscoveryService` (services-api).
- Rename `PublicIpDiscoveryServiceTest` → `PublicIpv4DiscoveryServiceImplTest`;
  move from `core` test sources to `services/ip` test sources.
- `LocalSipHostProvider` now injects `PublicIpv4DiscoveryService`
  (the interface) instead of the concrete class.
- Remove `requires java.ws.rs`, JAX-RS API dep, JAXB API dep, and CXF test
  dep from `core` — no remaining JAX-RS usage there.
